refactor(#17): move template bot into bot package
